Rename DeleteReplyConnection to UpdateReplyConnectionStatus
